Fix tableview header date label
I fixed an error where the header of the transaction tableview would not show “Yesterday” if the transaction occured less than 24 hours from the present time, even if it was a different calendar date.  I fixed it by removing the hour and minute of a transaction when the date is being compared in order to return the sectionHeader text.

bump to 1.3.4 b15